Add SSL flavors
Both flavors do exactly the same thing, with one difference: what port
they choose to expose in --discover and --connection-url.

For the SSL flavor, that's port 6380 where stunnel is listening over
SSL. For the "regular" flavor, that's port 6379 where bare Redis is
listening.

Once we add support for multiple ports in databases, we can simply
expose both these ports and let the customer choose which one to connect
to.