Work-around flaky test
This sets `test_after` then immediately attempts to read it. That may or
may not not, and we should give Redis a little time to propagate the
value.

For now, this just sleeps, which is fairly easy and does the trick, but
if we need more sleeps, we should consider implementing a busy wait.