Bump stunnel to 5.44
We've had a customer run into an issue connecting to Redis using the
Ruby client. While there is no obvious smoking gun in any of the logs,
stunnel did log a number of possible memory leaks and SSL errors,
mainly instances of:

- SSL_accept: Peer suddenly disconnected
- SSL_read: 1408F119: error:1408F119:SSL routines:SSL3_GET_RECORD:decryption failed or bad record mac

Unfortunately, from the logs (which is all we have at this point), it's
unclear if the memory leak caused SSL errors, if the errors caused the
memory leaks, or if the two are both consequences of another problem
(e.g. a TLS client misbehaving).

That said, considering Stunnel has a few allocation-related and
OpenSSL-related bugfixes between 5.37 and 5.44, updating is probably a
good idea here.

See: https://aptible.zendesk.com/agent/tickets/13316